+++ webinterface/analysis/analysis.py
@@ -251,7 +251,7 @@ def set_timeline_flows(profile, timewindow):
 
 @analysis.route("/timeline/<profile>/<timewindow>")
 # @analysis.route("/timeline/<profile>/<timewindow>/<search_filter>")
-def set_timeline(profile, timewindow, search_filter=""):
+def set_timeline(profile, timewindow,):
     """
     Set timeline data of a chosen profile and timewindow
     :return: list of timeline as set initially in database
@@ -260,36 +260,26 @@ def set_timeline(profile, timewindow, search_filter=""):
 
     timeline = __database__.db.zrange("profile_" + profile + "_" + timewindow + "_timeline", 0, -1)
     if timeline:
-
-        search_filter = search_filter.strip()
-        search = True if search_filter else False
-        reverse = False
-        if search and "!" in search_filter:
-            search_filter = search_filter.replace("!", "")
-            reverse = True
-
         for flow in timeline:
             flow = json.loads(flow)
+
+            # TODO: check IGMP
+            if flow["dport_name"] == "IGMP":
+                flow["dns_resolution"] = "????"
+                flow["dport/proto"] = "????"
+                flow["state"] = "????"
+                flow["sent"] = "????"
+                flow["recv"] = "????"
+                flow["tot"] = "????"
+                flow["warning"] = "????"
+                flow["critical warning"] = "????"
 
             # TODO: check this logic
             if flow["preposition"] == "from":
                 temp = flow["saddr"]
                 flow["daddr"] = temp
 
-            # search
-            if not search:
-                data.append(flow)
-            else:
-                value_is_present = False
-
-                # partial search in each flow key
-                for v in flow.values():
-                    if search_filter.lower() in str(v).lower():
-                        value_is_present = True
-                        break
-
-                if (not reverse and value_is_present) or (reverse and not value_is_present):
-                    data.append(flow)
+            data.append(flow)
 
     return {
         'data': data