+++ webinterface/static/app.js
@@ -28,17 +28,20 @@ $("#table_choose_redis").DataTable({
     ]
 })
 
-
 $('#modal_choose_redis').modal({
     show: false,
     backdrop: 'static',
     keyboard: false
 })
 
+$('#modal_choose_redis').on('show.bs.modal', function (e) {
+    $('#table_choose_redis').DataTable().ajax.reload();
+})
+
 $('#button_choose_db').click(function(){
-    let data = $('#table_choose_redis').DataTable().row( { selected: true } ).data()
+    let chosen_db = $('#table_choose_redis').DataTable().row( { selected: true } ).data()
     $('#modal_choose_redis .close').click() // close modal by imitating the close button click. $('#myModal').hide() does not work
-    let link = "/db/" + data['redis_port']
+    let link = "/db/" + chosen_db['redis_port']
     $.get( link );
     window.location.reload();
 });